notebook error action added and notebook_errors reducer file added

+++ frontend/components/notebooks/notebooks_index.jsx
@@ -26,10 +26,10 @@ class NotebooksIndex extends React.Component{
     }
 
     handleDeleteNotebook(notebookId) {
-        this.props.deleteNotebook(notebookId);
-        // if (this.state.title !== "Default") {
-        //     this.props.deleteNotebook(notebookId);
-        // }
+        // this.props.deleteNotebook(notebookId);
+        if (this.state.title !== "Default") {
+            this.props.deleteNotebook(notebookId);
+        }
     }
 
     handleCreateNotebook() {
